# apps/server/internal/api/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Robot Center 관제 서버 API입니다. Robot, Operator, Recorder, System 역할별 API를 제공합니다.",
        "title": "Robot Center 관제 API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/operator/sensor-descriptors": {
            "get": {
                "description": "missionId, robotCode 조건으로 센서 descriptor 목록을 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "센서 descriptor 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mission ID 또는 missionCode",
                        "name": "missionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "로봇 코드",
                        "name": "robotCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorDescriptorsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/operator/sensor-latest": {
            "get": {
                "description": "missionId, robotCode 조건으로 센서별 최신 sample을 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "센서 latest 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mission ID 또는 missionCode",
                        "name": "missionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "로봇 코드",
                        "name": "robotCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorLatestListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/operator/sensor-samples": {
            "get": {
                "description": "missionId, robotCode, sensorId 조건과 limit으로 센서 sample 목록을 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "센서 sample 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mission ID 또는 missionCode",
                        "name": "missionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "로봇 코드",
                        "name": "robotCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "센서 ID",
                        "name": "sensorId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "조회 개수 제한",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorSamplesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/operator/sfu/ws": {
            "get": {
                "description": "관제 UI operator peer가 mission room에 WebRTC signaling용 WebSocket으로 접속합니다.",
                "tags": [
                    "Operator API"
                ],
                "summary": "Operator SFU WebSocket 연결",
                "parameters": [
                    {
                        "type": "string",
                        "description": "접속할 missionCode room",
                        "name": "room",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/recorder/sensor-samples": {
            "post": {
                "description": "recorder-worker가 DataChannel sensor envelope를 저장합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "센서 sample 저장",
                "parameters": [
                    {
                        "description": "센서 envelope",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorEnvelopeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorSamplesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/recorder/sfu/ws": {
            "get": {
                "description": "recorder-worker peer가 mission room에 WebRTC signaling용 WebSocket으로 접속합니다.",
                "tags": [
                    "Recorder API"
                ],
                "summary": "Recorder SFU WebSocket 연결",
                "parameters": [
                    {
                        "type": "string",
                        "description": "접속할 missionCode room",
                        "name": "room",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/robot/sfu/ws": {
            "get": {
                "security": [
                    {
                        "RobotBearerAuth": []
                    }
                ],
                "description": "Robot publisher가 mission room에 WebRTC signaling용 WebSocket으로 접속합니다.",
                "tags": [
                    "Robot API"
                ],
                "summary": "Robot SFU WebSocket 연결",
                "parameters": [
                    {
                        "type": "string",
                        "description": "접속할 missionCode room",
                        "name": "room",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "robot-center_apps_server_internal_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.SensorDescriptorRequest": {
            "type": "object",
            "properties": {
                "channelRole": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                },
                "sensorId": {
                    "type": "string"
                },
                "sensorType": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorDescriptorResponse": {
            "type": "object",
            "properties": {
                "channelRole": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "firstSeenAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "lastSeenAt": {
                    "type": "string"
                },
                "missionId": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "sensorId": {
                    "type": "string"
                },
                "sensorType": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorDescriptorsResponse": {
            "type": "object",
            "properties": {
                "sensorDescriptors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorDescriptorResponse"
                    }
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorEnvelopeRequest": {
            "type": "object",
            "properties": {
                "channelRole": {
                    "type": "string"
                },
                "descriptors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorDescriptorRequest"
                    }
                },
                "messageId": {
                    "type": "string"
                },
                "messageType": {
                    "type": "string"
                },
                "missionId": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "samples": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorSampleRequest"
                    }
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorLatestListResponse": {
            "type": "object",
            "properties": {
                "missionId": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "sensors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorLatestResponse"
                    }
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorLatestResponse": {
            "type": "object",
            "properties": {
                "channelRole": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "firstSeenAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "lastSeenAt": {
                    "type": "string"
                },
                "latestSample": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorSampleResponse"
                },
                "missionId": {
                    "type": "string"
                },
                "readings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorValueReadingResponse"
                    }
                },
                "robotCode": {
                    "type": "string"
                },
                "sensorId": {
                    "type": "string"
                },
                "sensorType": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorSampleRequest": {
            "type": "object",
            "properties": {
                "channelRole": {
                    "type": "string"
                },
                "messageId": {
                    "type": "string"
                },
                "objectKey": {
                    "type": "string"
                },
                "sensorId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "values": {}
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorSampleResponse": {
            "type": "object",
            "properties": {
                "channelRole": {
                    "type": "string"
                },
                "descriptorId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messageId": {
                    "type": "string"
                },
                "missionId": {
                    "type": "string"
                },
                "objectKey": {
                    "type": "string"
                },
                "receivedAt": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "sensorId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorSamplesResponse": {
            "type": "object",
            "properties": {
                "sensorSamples": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.SensorSampleResponse"
                    }
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.SensorValueReadingResponse": {
            "type": "object",
            "properties": {
                "fieldPath": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "order": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
    },
    "securityDefinitions": {
        "RobotBearerAuth": {
            "description": "Robot API에서 사용하는 Bearer robotToken입니다.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}